requirements.txt
customtkinter==5.2.0
darkdetect==0.7.1
qrcode==7.4.2
Pillow==10.0.0
printer==2.0.0
win32printing==0.1.3
openpyxl==3.1.2
tkcalendar==1.6.1
```